Decouple coordinator from entities.
